TEST Mise a jour des tests pour FormAddArticle
