+++ __tests__/stockArticlesTitle.test.ts
@@ -1,6 +1,18 @@
-import { getStockArticlesTitle } from '../src/utils/Utils';
-import test from 'node:test';
+import getStockArticlesTitle  from '../src/utils/getStockArticlesTitle';
 import { describe, expect, it } from 'vitest';
+
+const fakeProductTest = {
+  id: 123,
+  title: "produit",
+  price: 5,
+  description: 'faux produit pour test',
+  category: {
+    id: 1,
+    name: "category",
+    image: "url"
+  },
+  images: ["url"]
+};
 
 describe('function getStockArticlesTitle" ', () => {
   it("getStockArticlesTitle([]) doit retourner 'aucun article disponible'", () => {
@@ -8,11 +20,11 @@ describe('function getStockArticlesTitle" ', () => {
     expect(messageArticle).toBe('Aucun article disponible');
   });
   it("getStockArticlesTitle([{},{}]) doit retourner '2 articles disponibles'", () => {
-    const messageArticle = getStockArticlesTitle([{}, {}]);
+    const messageArticle = getStockArticlesTitle([fakeProductTest, fakeProductTest]);
     expect(messageArticle).toBe('2 articles disponibles');
   });
   it("getStockArticlesTitle([{}]) doit retourner '1 article disponible'", () => {
-    const messageArticle = getStockArticlesTitle([{}]);
+    const messageArticle = getStockArticlesTitle([fakeProductTest]);
     expect(messageArticle).toBe('Un article disponible');
   });
 });
